--- a/cmd/generate/generate.go
+++ b/cmd/generate/generate.go
@@ -41,7 +41,7 @@ func NewCmdGenerate() *cobra.Command {
 // should be renamed appropriately, as of now it made no sense to generically
 // state pools when other pools aren't supported.
 func NewCmdGenerateCStorStoragePoolCluster() *cobra.Command {
-	var nodes, raidType string
+	var nodes, raidType, cap string
 	var devices int
 	cmd := &cobra.Command{
 		Use:   "cspc",
@@ -49,15 +49,18 @@ func NewCmdGenerateCStorStoragePoolCluster() *cobra.Command {
 		Run: func(cmd *cobra.Command, args []string) {
 			node, _ := cmd.Flags().GetString("nodes")
 			raid, _ := cmd.Flags().GetString("raidtype")
+			capacity, _ := cmd.Flags().GetString("capacity")
 			devs := numDevices(cmd)
 			nodeList := strings.Split(node, ",")
-			util.CheckErr(generate.CSPC(nodeList, devs, raid), util.Fatal)
+			util.CheckErr(generate.CSPC(nodeList, devs, raid, capacity), util.Fatal)
 		},
 	}
 	cmd.PersistentFlags().StringVarP(&nodes, "nodes", "", "",
 		"comma separated set of nodes for pool creation --nodes=node1,node2,node3,node4")
 	_ = cmd.MarkPersistentFlagRequired("nodes")
 	cmd.PersistentFlags().StringVarP(&raidType, "raidtype", "", "stripe",
+		"allowed RAID configuration such as, stripe, mirror, raid, raidz2")
+	cmd.PersistentFlags().StringVarP(&cap, "capacity", "", "10Gi",
 		"allowed RAID configuration such as, stripe, mirror, raid, raidz2")
 	cmd.PersistentFlags().IntVar(&devices, "number-of-devices", 1, "number of devices per node, selects default based on raid-type")
 	return cmd