--- a/pkg/generate/cspc.go
+++ b/pkg/generate/cspc.go
@@ -26,6 +26,7 @@ import (
 	"github.com/openebs/api/v2/pkg/apis/openebs.io/v1alpha1"
 	"github.com/openebs/openebsctl/pkg/client"
 	"github.com/openebs/openebsctl/pkg/util"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -38,13 +39,20 @@ func isPoolTypeValid(raid string) bool {
 }
 
 // CSPC calls the generate routine for different cas-types
-func CSPC(nodes []string, devs int, raid string) error {
+func CSPC(nodes []string, devs int, raid, capacity string) error {
 	c := client.NewK8sClient()
 	if !isPoolTypeValid(strings.ToLower(raid)) {
 		// TODO: Use the well defined pool constant types from openebs/api when added there
 		return fmt.Errorf("invalid pool type %s", raid)
 	}
-	_, str, err := cspc(c, nodes, devs, strings.ToLower(raid))
+	// resource.Quantity doesn't like the bits or bytes suffixes
+	capacity = strings.Replace(capacity, "b", "", 1)
+	capacity = strings.Replace(capacity, "B", "", 1)
+	size, err := resource.ParseQuantity(capacity)
+	if err != nil {
+		return err
+	}
+	_, str, err := cspc(c, nodes, devs, strings.ToLower(raid), size)
 	if err != nil {
 		return err
 	}
@@ -53,13 +61,23 @@ func CSPC(nodes []string, devs int, raid string) error {
 }
 
 // cspc takes eligible nodes, number of devices and poolType to create a pool cluster template
-func cspc(c *client.K8sClient, nodes []string, devs int, poolType string) (*cstorv1.CStorPoolCluster, string, error) {
+func cspc(c *client.K8sClient, nodes []string, devs int, poolType string, minSize resource.Quantity) (*cstorv1.CStorPoolCluster, string, error) {
 	// 0. Figure out the OPENEBS_NAMESPACE for CStor
 	cstorNS, err := c.GetOpenEBSNamespace(util.CstorCasType)
 	// assume CSTOR's OPENEBS_NAMESPACE has all the relevant blockdevices
 	c.Ns = cstorNS
 	if err != nil {
 		return nil, "", fmt.Errorf("unable to determine the cStor namespace error: %v", err)
+	}
+	// 0.1 Validate user input, check if user hasn't entered less than 64Mi
+	cstorMin := resource.MustParse("64Mi")
+	if minSize.Cmp(cstorMin) < 0 {
+		return nil, "", fmt.Errorf("minimum size of supported block-devices in a cspc is 64Mi")
+	}
+	// 0.2 Validate user input, check if user has entered >= minimum supported BD-count
+	if min := minCount()[poolType]; devs < min {
+		return nil, "", fmt.Errorf("%s pool requires a minimum of %d block device per node",
+			poolType, min)
 	}
 	// 1. Validate nodes & poolType, fetch disks
 	nodeList, err := c.GetNodes(nodes, "", "")
@@ -82,7 +100,7 @@ func cspc(c *client.K8sClient, nodes []string, devs int, poolType string) (*csto
 	if err != nil || len(bds.Items) == 0 {
 		return nil, "", fmt.Errorf("no blockdevices found in nodes with %v hostnames", hostnames)
 	}
-	_, err = filterCStorCompatible(bds)
+	_, err = filterCStorCompatible(bds, minSize)
 	if err != nil {
 		return nil, "", fmt.Errorf("(server error) unable to fetch bds from %v nodes", nodes)
 	}
@@ -183,13 +201,11 @@ func makePools(poolType string, nDevices int, bd map[string][]v1alpha1.BlockDevi
 		}
 		return &spec, nil
 	case string(cstorv1.PoolMirrored), string(cstorv1.PoolRaidz), string(cstorv1.PoolRaidz2):
-		// 1. Check for devices per node count against the fixed minimum
 		min := minCount()[poolType]
-		if nDevices < min {
-			return nil, fmt.Errorf("%s pool requires a minimum of %d block device per node",
-				poolType, min)
+		if min > nDevices {
+			return nil, fmt.Errorf("insufficient blockdevices expected for %s", poolType)
 		}
-		// 2. Start filling in the devices in their RAID-groups per the hostnames
+		// 1. Start filling in the devices in their RAID-groups per the hostnames
 		for i, host := range hosts {
 			var raidGroups []cstorv1.RaidGroup
 			// add all BDs to a CSPCs CSPI spec
@@ -198,9 +214,14 @@ func makePools(poolType string, nDevices int, bd map[string][]v1alpha1.BlockDevi
 				return nil, fmt.Errorf("not enough eligible blockdevices found on node %s, want %d, found %d", nodes[i], nDevices, len(bds))
 			}
 			index := 0
+			maxIndex := len(bds)
+			if maxIndex < nDevices {
+				return nil, fmt.Errorf("not enough eligible blockdevices found on node %s, want %d, found %d", nodes[i], min, maxIndex)
+			}
 			for d := 0; d < nDevices/min; d++ {
 				var raids []cstorv1.CStorPoolInstanceBlockDevice
 				for j := 0; j < min; j++ {
+					// each RaidGroup has min number of devices
 					raids = append(raids, cstorv1.CStorPoolInstanceBlockDevice{BlockDeviceName: bds[index].Name})
 					index++
 				}
@@ -224,6 +245,7 @@ func makePools(poolType string, nDevices int, bd map[string][]v1alpha1.BlockDevi
 // minCount states the minimum number of BDs for a pool type
 func minCount() map[string]int {
 	return map[string]int{
+		string(cstorv1.PoolStriped):  1,
 		string(cstorv1.PoolMirrored): 2,
 		string(cstorv1.PoolRaidz):    3,
 		string(cstorv1.PoolRaidz2):   4,
@@ -231,7 +253,7 @@ func minCount() map[string]int {
 }
 
 // filterCStorCompatible takes a list of BDs and gives out a list of BDs which can be used to provision a pool
-func filterCStorCompatible(bds *v1alpha1.BlockDeviceList) (*v1alpha1.BlockDeviceList, error) {
+func filterCStorCompatible(bds *v1alpha1.BlockDeviceList, minLimit resource.Quantity) (*v1alpha1.BlockDeviceList, error) {
 	// TODO: Optionally reject sparse-disks depending on configs
 	var final []v1alpha1.BlockDevice
 	for _, bd := range bds.Items {
@@ -240,13 +262,13 @@ func filterCStorCompatible(bds *v1alpha1.BlockDeviceList) (*v1alpha1.BlockDevice
 			bd.Status.ClaimState == v1alpha1.BlockDeviceUnclaimed &&
 			bd.Spec.FileSystem.Type == "" &&
 			// BD's capacity >=64 MiB
-			bd.Spec.Capacity.Storage >= 67110000 {
+			bd.Spec.Capacity.Storage >= uint64(minLimit.Value()) {
 			final = append(final, bd)
 		}
 	}
 	bds.Items = final
 	if len(final) == 0 {
-		return nil, fmt.Errorf("found no eligble blockdevices")
+		return nil, fmt.Errorf("found no eligble blockdevices of size %s", minLimit.String())
 	}
 	return bds, nil
 }