--- a/pkg/generate/cspc.go
+++ b/pkg/generate/cspc.go
@@ -232,10 +232,10 @@ func makePools(poolType string, nDevices int, bd map[string][]v1alpha1.BlockDevi
 			devices := Generate(v1alpha1.BlockDeviceList{Items: bds})
 			for d := 0; d < nDevices/min; d++ {
 				var raids []cstorv1.CStorPoolInstanceBlockDevice
-				tmp, thisRaidGroup, err := devices.Select(minsize, min)
+				d, thisRaidGroup, err := devices.Select(minsize, min)
 				// re-assign the head node of the linked-list for next iteration
-				//
-				devices = tmp
+				// pinning the new head to the variable declared above for upcoming usage as required
+				devices = d
 				if err != nil {
 					return nil, err
 				}