As of now, the minimum number of Blockdevices get added into a single RAIDgroup.

TODO: ~Consider storage sizes for creating a pool~.

# Implementation to minimize disk losses

- For stripe, there can be only one Raid-Group(in a CSPI) where all the disks exist and they can be of any sizes
- For Mirror, RaidZ, RaidZ2 the BD sizes must be the same in a single `CSPI.RaidGroup`( same size exact to the number of bytes, one byte off and the CSPC will not provision because of a ZFS error)

## How the BD selection works for non-stripe pools?

### Tl; dr summary

1. Filter BDs by minimumSize, Unclaimed+Active states
2. Sort them by increasing order of size
3. Assume, minimum device-level replication for mirror, raidZ, Raidz2(i.e. they'll always have `2, 3 and 6` devices per-RAIDgroup)
4. Select disks from the list^ of disks available in the node

### Step-4 in detail

1. Have a sorted slice of BDs
2. Convert the slice of BDs into a singly-linked-list
3. Run a two-pointer algorithm similar to Sliding Window, the first pointer(i.e. `curr`) will point to the start of the window and the second pointer(i.e. `ahead`) will point to the end of the window, including both the boundaries, they'd always have `1, 2, 3, 6` boxes in their bounds.
4. If the sizes of the BD pointed by the `curr` BD and the `ahead` BD are same, we can confirm that the n elements between them are of the same size, where `n ∀ n ∈ {0, 1, 2..}`
NOTE: If the `curr` & `ahead` point to the same BD node, the single BD is selected

5. We can get rid of those BD nodes by keeping track of the `prev` pointer to the BD pointed to by the `curr` BD
  - the edge cases around selection of the first BD is handled by prefixing a dummy useless node in the `Select(...)` method which is never shared with the caller in any of the edge cases
6. If a chosen minimum number of same sized BDs aren't found between the two pointers, we can confirm that a specific node doesn't have enough BDs for that RaidGroup in a particular pool-instance and promptly throw an error

## Future possible enhancements

- Add a feature flag for omitting HDDs, sparse-files, selecting only HDDs, etc by just modifying `filterCStorCompatibleBDs` function
- Add some automatically fuzz-ed unit tests to further harden the logic


Co-Authored-by: Sai Chaithanya (@mittachaitu) <sai.chaithanya@mayadata.io>
Signed-off-by: Harsh Vardhan (@vharsh) <harsh.vardhan@mayadata.io>