@vharsh 2021-12-24
Makes sense, I’ll have to change how unit tests check them internally, but
this seems doable.

On Fri, 24 Dec 2021 at 1:17 PM, sai chaithanya ***@***.***>
wrote:

> ***@***.**** approved this pull request.
>
> LGTM
> ------------------------------
>
> In pkg/generate/sort.go
> <https://github.com/openebs/openebsctl/pull/136#discussion_r774914482>:
>
> > @@ -0,0 +1,86 @@
> +package generate
> +
> +import (
> +	"fmt"
> +
> +	"github.com/openebs/api/v2/pkg/apis/openebs.io/v1alpha1"
> +	"k8s.io/apimachinery/pkg/api/resource"
> +)
> +
> +// DeviceList is a LinkedList of BlockDevices
> +type DeviceList struct {
> +	item v1alpha1.BlockDevice
>
> nit: If we made it as pointer copying of things will be reduced to great
> extent as compared to struct copy
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/openebs/openebsctl/pull/136#pullrequestreview-839789164>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACG7P4XKZYHES6UYJTURNB3USQQQ5ANCNFSM5JEF3M7Q>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>