@vharsh 2021-12-24
> reflect.DeepEqual should work

Yes, it did, made some minor changes in the unit test data. It likely also improved performance numbers.