@vharsh 2021-12-24
Before:
```
BenchmarkSelect
BenchmarkSelect/six_node_LinkedList,_two_BD_required_of_6G
BenchmarkSelect/six_node_LinkedList,_two_BD_required_of_6G-8         	  282495	      4133 ns/op
BenchmarkSelect/six_node_LinkedList_with_unsorted_BD_sizes,_two_BD_required_of_1G
BenchmarkSelect/six_node_LinkedList_with_unsorted_BD_sizes,_two_BD_required_of_1G-8         	  232467	      5830 ns/op
PASS
```


After:
```
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
BenchmarkSelect/six_node_LinkedList,_two_BD_required_of_6G-8         	  414178	      3910 ns/op	     848 B/op	      15 allocs/op
BenchmarkSelect/six_node_LinkedList_with_unsorted_BD_sizes,_two_BD_required_of_1G-8         	  282440	      5336 ns/op	     944 B/op	      23 allocs/op
PASS
ok  	github.com/openebs/openebsctl/pkg/generate	3.279s
```